menambah tabel comments, likes, dan reports

+++ database/migrations/2026_05_31_161932_create_comments_table.php
@@ -11,13 +11,15 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('comments', function (Blueprint $table) {
-    $table->id();
-    $table->foreignId('user_id')->constrained()->onDelete('cascade');
-    $table->text('body');
-    $table->morphs('commentable'); 
-    $table->timestamps();
-});
+        if (!Schema::hasTable('comments')) {
+            Schema::create('comments', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->text('body');
+                $table->morphs('commentable'); 
+                $table->timestamps();
+            });
+        }
     }
 
     /**
