+++ database/migrations/2026_05_31_161939_create_likes_table.php
@@ -11,13 +11,15 @@ return new class extends Migration
      */
     public function up(): void
     {
-       Schema::create('likes', function (Blueprint $table) {
-    $table->id();
-    $table->foreignId('user_id')->constrained()->onDelete('cascade');
-    $table->morphs('likeable'); 
-    $table->timestamps();
-    $table->unique(['user_id', 'likeable_id', 'likeable_type']);
-});
+       if (!Schema::hasTable('likes')) {
+           Schema::create('likes', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->morphs('likeable'); 
+                $table->timestamps();
+                $table->unique(['user_id', 'likeable_id', 'likeable_type']);
+            });
+       }
     }
 
     /**
