+++ database/migrations/2026_05_31_162043_create_reports_table.php
@@ -11,14 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-       Schema::create('reports', function (Blueprint $table) {
-    $table->id();
-    $table->foreignId('user_id')->constrained()->onDelete('cascade');
-    $table->string('reason'); // Alasan report (e.g., spam, hate speech)
-    $table->text('description')->nullable();
-    $table->morphs('reportable'); 
-    $table->timestamps();
-});
+       if (!Schema::hasTable('reports')) {
+           Schema::create('reports', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->string('reason'); // Alasan report (e.g., spam, hate speech)
+                $table->text('description')->nullable();
+                $table->morphs('reportable'); 
+                $table->timestamps();
+            });
+       }
     }
 
     /**
